Move Node.js Lambda treeshaking into its own dir, stub in Python
