+++ src/actions/autoinstall/node/get-root-deps.js
